+++ Courier/src/utopia/courier/controller/read/LazyEmailHeadersView.scala
@@ -1,7 +1,7 @@
 package utopia.courier.controller.read
 
+import utopia.courier.model.EmailHeaders
 import utopia.courier.model.write.Recipients
-import utopia.courier.model.{EmailHeaders, EmailHeadersLike}
 import utopia.flow.time.Now
 import utopia.flow.util.StringExtensions._
 
@@ -12,10 +12,11 @@ import java.time.Instant
   * @author Mikko Hilpinen
   * @since 13.9.2021, v0.1
   */
-class LazyEmailHeadersView(getSender: => String, getSubject: => String, getSendTime: => Instant,
-                           getRecipients: => Recipients = Recipients.empty, geReplyTo: => String = "",
+class LazyEmailHeadersView(getSender: => String, getSubject: => String, getMessageId: => String, getSendTime: => Instant,
+                           getRecipients: => Recipients = Recipients.empty, getInReplyTo: => String,
+                           getReferences: => Vector[String], geReplyTo: => String = "",
                            override val receiveTime: Instant = Now)
-	extends EmailHeadersLike
+	extends EmailHeaders
 {
 	// ATTRIBUTES   ---------------------------------
 	
@@ -24,6 +25,9 @@ class LazyEmailHeadersView(getSender: => String, getSubject: => String, getSendT
 	override lazy val subject = getSubject
 	override lazy val replyTo = geReplyTo.notEmpty.getOrElse(sender)
 	override lazy val sendTime = getSendTime
+	override lazy val messageId: String = getMessageId
+	override lazy val inReplyTo: String = getInReplyTo
+	override lazy val references: Vector[String] = getReferences
 	
 	
 	// COMPUTED ------------------------------------
@@ -31,5 +35,6 @@ class LazyEmailHeadersView(getSender: => String, getSubject: => String, getSendT
 	/**
 	  * @return A fully cached / read headers instance based on this view
 	  */
-	def toHeaders = EmailHeaders.incoming(sender, subject, sendTime, recipients, replyTo, receiveTime)
+	def toHeaders = EmailHeaders.incoming(sender, subject, messageId, sendTime, recipients, inReplyTo, references,
+		replyTo, receiveTime)
 }